Add files via upload
each function that should returns a single dict of that entity. also update remote no longer requires $push to be in there, it inserts it itself so it can keep the update data clean. Kept this as mongodb2 so I can look at the differences if I need to reset it.